+++ examples/beat_app.rs
@@ -1,9 +1,9 @@
 #![allow(unused_variables)]
 
+use anyhow::Result;
+use celery::beat::RegularSchedule;
 use celery::task::TaskResult;
-use celery::RegularSchedule;
 use env_logger::Env;
-use exitfailure::ExitFailure;
 use tokio::time::Duration;
 
 const QUEUE_NAME: &str = "celery";
@@ -19,7 +19,7 @@ fn long_running_task(secs: Option<u64>) -> TaskResult<()> {
 }
 
 #[tokio::main]
-async fn main() -> Result<(), ExitFailure> {
+async fn main() -> Result<()> {
     env_logger::from_env(Env::default().default_filter_or("info")).init();
 
     // Build a `Beat` with a default scheduler backend.