+++ src/task/signature.rs
@@ -1,4 +1,4 @@
-use super::Task;
+use super::{Task, TaskOptions};
 use chrono::{DateTime, Utc};
 
 /// Wraps the parameters and execution options for a single task invocation.
@@ -9,7 +9,7 @@ use chrono::{DateTime, Utc};
 /// # Examples
 ///
 /// ```rust
-/// use celery::TaskResult;
+/// use celery::prelude::*;
 ///
 /// #[celery::task]
 /// fn add(x: i32, y: i32) -> TaskResult<i32> {
@@ -17,8 +17,6 @@ use chrono::{DateTime, Utc};
 /// }
 ///
 /// let signature = add::new(1, 2);
-/// assert_eq!(signature.params.x, 1);
-/// assert_eq!(signature.params.y, 2);
 /// ```
 #[derive(Clone)]
 pub struct Signature<T>
@@ -26,40 +24,30 @@ where
     T: Task,
 {
     /// The parameters for the task invocation.
-    pub params: T::Params,
+    pub(crate) params: T::Params,
 
     /// A queue to send the task to.
-    pub queue: Option<String>,
-
-    /// Time limit, in seconds, for the task execution. Overrides any app or task-level default time limits.
-    pub time_limit: Option<u32>,
-
-    /// The `time_limit` option is equivalent to the ["soft time
-    /// limit"](https://docs.celeryproject.org/en/stable/userguide/workers.html#time-limits)
-    /// option when sending tasks to a Python consumer.
-    /// If you desire to set a "hard time limit", use this option.
-    ///
-    /// *Note that this is really only for compatability with Python workers*.
-    /// `time_limit` and `hard_time_limit` are treated the same by Rust workers, and if both
-    /// are set, the minimum of the two will be used.
-    pub hard_time_limit: Option<u32>,
+    pub(crate) queue: Option<String>,
 
     /// The number of seconds to wait before executing the task. This is equivalent to setting
     /// [`eta`](struct.Signature.html#structfield.eta)
     /// to `current_time + countdown`.
-    pub countdown: Option<u32>,
+    pub(crate) countdown: Option<u32>,
 
     /// A future ETA at which to execute the task.
-    pub eta: Option<DateTime<Utc>>,
+    pub(crate) eta: Option<DateTime<Utc>>,
 
     /// A number of seconds until the task expires, at which point it should no longer
     /// be executed. This is equivalent to setting
     /// [`expires`](struct.Signature.html#structfield.expires)
     /// to `current_time + expires_in`.
-    pub expires_in: Option<u32>,
+    pub(crate) expires_in: Option<u32>,
 
     /// A future time at which the task will expire.
-    pub expires: Option<DateTime<Utc>>,
+    pub(crate) expires: Option<DateTime<Utc>>,
+
+    /// Additional options.
+    pub(crate) options: TaskOptions,
 }
 
 impl<T> Signature<T>
@@ -71,12 +59,11 @@ where
         Self {
             params,
             queue: None,
-            time_limit: None,
-            hard_time_limit: None,
             countdown: None,
             eta: None,
             expires_in: None,
             expires: None,
+            options: TaskOptions::default(),
         }
     }
 
@@ -88,22 +75,6 @@ where
     /// Set the queue.
     pub fn with_queue(mut self, queue: &str) -> Self {
         self.queue = Some(queue.into());
-        self
-    }
-
-    /// Set a time limit (in seconds) for the task.
-    pub fn with_time_limit(mut self, time_limit: u32) -> Self {
-        self.time_limit = Some(time_limit);
-        self
-    }
-
-    /// Set a hard time limit (in seconds) for the task.
-    ///
-    /// *Note that this is really only for compatability with Python workers*.
-    /// `time_limit` and `hard_time_limit` are treated the same by Rust workers, and if both
-    /// are set, the minimum of the two will be used.
-    pub fn with_hard_time_limit(mut self, time_limit: u32) -> Self {
-        self.hard_time_limit = Some(time_limit);
         self
     }
 
@@ -128,6 +99,22 @@ where
     /// Set the expiration time.
     pub fn with_expires(mut self, expires: DateTime<Utc>) -> Self {
         self.expires = Some(expires);
+        self
+    }
+
+    /// Set a time limit (in seconds) for the task.
+    pub fn with_time_limit(mut self, time_limit: u32) -> Self {
+        self.options.time_limit = Some(time_limit);
+        self
+    }
+
+    /// Set a hard time limit (in seconds) for the task.
+    ///
+    /// *Note that this is really only for compatability with Python workers*.
+    /// `time_limit` and `hard_time_limit` are treated the same by Rust workers, and if both
+    /// are set, the minimum of the two will be used.
+    pub fn with_hard_time_limit(mut self, time_limit: u32) -> Self {
+        self.options.hard_time_limit = Some(time_limit);
         self
     }
 }